Add another test and fix issue it found
Macro sub-parsers shouldn't try to resolve symbols.

Allow parsers to run just the `parse` method; provide a new `assemble`
method that does the full parse-and-link flow.
